+++ trunk/vstgui/uidescription/editing/uifontscontroller.cpp
@@ -280,7 +280,7 @@ void UIFontsController::valueChanged (CControl* pControl)
 //----------------------------------------------------------------------------------------------------
 void UIFontsController::dbSelectionChanged (int32_t selectedRow, GenericStringListDataBrowserSource* source)
 {
-	selectedFont = selectedRow != CDataBrowser::kNoSelection ? dataSource->getStringList ()->at (selectedRow).c_str () : "";
+	selectedFont = selectedRow != CDataBrowser::kNoSelection ? dataSource->getStringList ()->at (static_cast<uint32_t> (selectedRow)).c_str () : "";
 	CFontRef font = editDescription->getFont (selectedFont.c_str ());
 	if (font)
 	{
